feat(CodeEditor): enhance formatting capabilities and modularize code
- Introduced a new `formatters.ts` file to centralize formatting logic for various languages, improving code organization and maintainability.
- Added `useFormattingModuleLoaders` hook to dynamically load formatting modules based on the selected language, enhancing flexibility.
- Updated `useCodeFormatter` to utilize the new formatting functions and options, ensuring consistent behavior across different languages.
- Removed the deprecated `useCodeFormatter` implementation, streamlining the codebase and reducing redundancy.
- Added comprehensive tests for the new formatting logic, ensuring correct behavior across supported languages and handling of edge cases.
