+++ packages/code-editor/src/CodeEditor/hooks/formatting/index.ts
@@ -1,0 +1,20 @@
+// Main hooks
+export { useCodeFormatter } from './useCodeFormatter';
+export { useFormattingModuleLoaders } from './useFormattingModuleLoaders';
+
+// Core formatting functions
+export { formatCode } from './formatters';
+
+// Helper functions
+export {
+  createClangFormatConfig,
+  createJavaScriptConfig,
+  createPrettierConfig,
+} from './utils';
+
+// Types
+export type {
+  FormatCodeParams,
+  FormatCodeResult,
+  FormattingOptions,
+} from './types';
