+++ packages/code-editor/src/CodeEditor/hooks/formatting/useCodeFormatter.ts
@@ -1,0 +1,64 @@
+import { useEffect, useState } from 'react';
+
+import { type CodeEditorProps } from '../../CodeEditor.types';
+import { CodeEditorModules } from '../moduleLoaders.types';
+
+import { formatCode } from './formatters';
+import { FormattingOptions } from './types';
+import { areModulesLoaded } from './utils';
+
+/**
+ * Hook for creating a code formatting utility that can format code based on the selected language.
+ *
+ * This hook provides language-specific code formatting capabilities using lazy-loaded
+ * formatting modules. It supports both Prettier-based formatters and WASM-based formatters.
+ *
+ * @param params - Configuration object
+ * @param params.props - Partial CodeEditor props containing language setting
+ * @param params.modules - Module dependencies containing formatting modules
+ * @returns Object containing the format function and availability check
+ */
+export function useCodeFormatter({
+  props,
+  modules,
+}: {
+  props: Partial<CodeEditorProps>;
+  modules: Partial<CodeEditorModules>;
+}) {
+  // Extract editor configuration for consistent formatting
+  const editorTabWidth = props.indentSize ?? 2;
+  const editorUseTabs = props.indentUnit === 'tab';
+  const [isFormattingAvailable, setIsFormattingReady] = useState(false);
+
+  /**
+   * Formats code using the appropriate formatter based on the selected language.
+   *
+   * @param code - The code string to format
+   * @param options - Optional formatting options
+   * @returns Promise resolving to formatted code, or original code if formatting fails/unavailable
+   */
+  const handleFormatCode = async (
+    code: string,
+    options: FormattingOptions = {},
+  ): Promise<string> => {
+    return formatCode(
+      code,
+      props.language,
+      options,
+      editorTabWidth,
+      editorUseTabs,
+      modules,
+    );
+  };
+
+  useEffect(() => {
+    setIsFormattingReady(
+      (props.language && areModulesLoaded(props.language, modules)) || false,
+    );
+  }, [modules, props]);
+
+  return {
+    formatCode: handleFormatCode,
+    isFormattingAvailable,
+  };
+}
