+++ packages/code-editor/src/CodeEditor/index.ts
@@ -12,6 +12,7 @@ export {
 } from './CodeEditor.types';
 export {
   type CodeEditorModules,
+  type FormattingOptions,
   LanguageName,
   useAutoCompleteExtension,
   useCodeFoldingExtension,
